move decodeString and parseContentType in utils && normalize encoding
