+++ lib/jsdom.js
@@ -10,7 +10,7 @@ const CookieJar = require("tough-cookie").CookieJar;
 const toFileUrl = require("./jsdom/utils").toFileUrl;
 const defineGetter = require("./jsdom/utils").defineGetter;
 const defineSetter = require("./jsdom/utils").defineSetter;
-const parseContentType = require("./jsdom/utils").parseContentType;
+const parseContentType = require("./jsdom/living/helpers/headers").parseContentType;
 const documentFeatures = require("./jsdom/browser/documentfeatures");
 const domToHtml = require("./jsdom/browser/domtohtml").domToHtml;
 const Window = require("./jsdom/browser/Window");
