+++ lib/jsdom/browser/resource-loader.js
@@ -2,8 +2,8 @@
 
 const resolveHref = require("../utils").resolveHref;
 const parseDataUrl = require("../utils").parseDataUrl;
-const parseContentType = require("../utils").parseContentType;
-const decodeString = require("../utils").decodeString;
+const parseContentType = require("../living/helpers/headers").parseContentType;
+const decodeString = require("../living/helpers/encoding").decodeString;
 const fs = require("fs");
 const request = require("request");
 const documentBaseURL = require("../living/helpers/document-base-url").documentBaseURL;
