+++ lib/jsdom/living/file-reader.js
@@ -3,8 +3,8 @@
 const DOMException = require("../web-idl/DOMException");
 const EventTarget = require("./generated/EventTarget");
 const addConstants = require("../utils").addConstants;
-const parseContentType = require("../utils").parseContentType;
-const decodeString = require("../utils").decodeString;
+const parseContentType = require("./helpers/headers").parseContentType;
+const decodeString = require("./helpers/encoding").decodeString;
 const blobSymbols = require("./blob-symbols");
 const querystring = require("querystring");
