+++ lib/jsdom/living/nodes/HTMLFrameElement-impl.js
@@ -5,7 +5,7 @@ const HTMLElementImpl = require("./HTMLElement-impl").implementation;
 const applyDocumentFeatures = require("../../browser/documentfeatures").applyDocumentFeatures;
 const resourceLoader = require("../../browser/resource-loader");
 const defineGetter = require("../../utils").defineGetter;
-const parseContentType = require("../../utils").parseContentType;
+const parseContentType = require("../helpers/headers").parseContentType;
 const URL = require("whatwg-url").URL;
 const documentBaseURL = require("../helpers/document-base-url").documentBaseURL;
 const getAttributeValue = require("../attributes").getAttributeValue;
