+++ lib/jsdom/living/xmlhttprequest.js
@@ -11,11 +11,11 @@ const xhrSymbols = require("./xmlhttprequest-symbols");
 const blobSymbols = require("./blob-symbols");
 const formDataSymbols = require("./form-data-symbols");
 const addConstants = require("../utils").addConstants;
-const parseContentType = require("../utils").parseContentType;
-const decodeString = require("../utils").decodeString;
-const tokenRegexp = require("../utils").tokenRegexp;
-const fieldValueRegexp = require("../utils").fieldValueRegexp;
-const headerListSeparatorRegexp = require("../utils").headerListSeparatorRegexp;
+const parseContentType = require("./helpers/headers").parseContentType;
+const decodeString = require("./helpers/encoding").decodeString;
+const tokenRegexp = require("./helpers/headers").tokenRegexp;
+const fieldValueRegexp = require("./helpers/headers").fieldValueRegexp;
+const headerListSeparatorRegexp = require("./helpers/headers").headerListSeparatorRegexp;
 const idlUtils = require("./generated/utils");
 const documentBaseURLHelper = require("./helpers/document-base-url");
 const Document = require("./generated/Document");
